requirements.txt
nltk>=3.8.1
transformers>=4.35.0
gensim>=4.3.1
scikit-learn>=1.3.2
scipy>=1.11.4
numpy>=1.24.3
torch>=2.1.0
pyyaml>=6.0.1
dotenv